Move Dashboard commands to a separate file
